refactor: move Athena-specific leveling code to lib/tools/athena/
Pure file move + import-path updates, no behavior change. The six moved
files all depend on athena_iot models or the Athena2 wizard config.
Generic screens (leveling_screen, manual_leveling_screen) stay in
lib/tools/.
