+++ lib/tools/leveling_screen.dart
@@ -20,8 +20,8 @@ import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:orion/backend_service/backend_registry.dart';
 import 'package:orion/backend_service/backend_service.dart';
 import 'package:orion/glasser/glasser.dart';
-import 'package:orion/tools/c3d_athena2_wizard.dart';
-import 'package:orion/tools/leveling_configs.dart';
+import 'package:orion/tools/athena/c3d_athena2_wizard.dart';
+import 'package:orion/tools/athena/leveling_configs.dart';
 import 'package:orion/tools/manual_leveling_screen.dart';
 import 'package:orion/util/orion_config.dart';
 import 'package:orion/util/orion_spacing.dart';
